% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_r}
\alias{eval_r}
\title{Evaluate Expression After Recursively Substituting It}
\usage{
eval_r(expr, envir = parent.frame(), enclos = if (is.list(envir) ||
  is.pairlist(envir)) parent.frame() else baseenv())
}
\arguments{
\item{expr}{an object to be evaluated, see \link{eval}.}
}
\value{
The result of evaluating the object: for an expression vector this is
the result of evaluating the last element.
}
\description{
Calls \link{recsub} on with the inputs, and then evaluates the result.
}
