% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recsub.R
\name{get_with_env}
\alias{get_with_env}
\title{Find Symbol and Environment it is Bound in}
\usage{
get_with_env(symb.chr, envir, mode = "any")
}
\arguments{
\item{symb.chr}{a character(1L) representation of symbol name}

\item{mode}{"any" or "function"}
}
\value{
a list with the object and environment it was found in if there was
one, NULL otherwise
}
\description{
Like \code{get}, except that it returns the symbol value and the environment it
was found in.
}
