% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recsub-package.R
\docType{package}
\name{oshka-package}
\alias{oshka-package}
\title{Programmable Non-Standard Evaluation}
\description{
Expands quoted language by recursively replacing any symbols that point to
language objects with that language.  Users can construct complex language
objects from simple ones, which they can then expand and evaluate.
}
