% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recsub-package.R
\docType{package}
\name{recsub-package}
\alias{recsub-package}
\title{Programmable Non-Standard Evaluation}
\description{
Provides tools that can be used implement programmable Non-Standard
Evaluation (NSE).  \code{recsub} expands symbols in quoted language that point to
language objects.  This allows users to store language in variables that can
then be automatically injected into other language expressions.
}
