% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.data}
\alias{state.data}
\title{State Data}
\format{An object of class \code{data.frame} with 50 rows and 6 columns.}
\usage{
state.data
}
\description{
Selected statistics from the R built-in \code{state} data set, as well as an
additional column indicating whether the state has at least one NFL team or
not.
}
\seealso{
\link{state}
}
\keyword{datasets}
