% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recsub-package.R
\docType{package}
\name{recsub-package}
\alias{recsub-package}
\title{Programmable Non-Standard Evaluation}
\description{
Tools to implement programmable Non-Standard Evaluation (NSE).  \code{recsub}
works by recursively substituting symbols that point to language objects.
The semantics are the same as they are for non-language R objects so usage is
intuitive.
}
