% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recsub.R
\name{recsub}
\alias{recsub}
\title{Recursively Substitute Language}
\usage{
recsub(expr, envir = parent.frame(), enclos = if (is.list(envir) ||
  is.pairlist(envir)) parent.frame() else baseenv())
}
\description{
Used to implement programmable Non-Standard Evaluation
}
\examples{
a <- quote(x > 3)
b <- quote(x < 10)
c <- quote(a & b)
recsub(c)
}
